Commit 6/2/2015
Better animation sprite, fixed a major bug where the sprite messed up
motion when given a movement command at a certain time. Added a queue to
animation sprite to keep track of movements. Added variable speed to
animations. Eliminated all superfluous comments and debugging
statements. stdSprite is still pretty much the same. Working on the
mechanics of showing and hiding a sprite. Once finished with that, I
will be finished with the animation sprite and will be ready to move on
to other things